requirements.txt
# --- Core detección ---
ultralytics>=8.3.193,<8.4

# --- UI / APIs ---
gradio==4.44.1
fastapi==0.111.0
uvicorn[standard]==0.30.5

# --- Visión / datos ---
opencv-python==4.9.0.80
pillow==10.4.0
numpy==2.0.2
pandas==2.2.2
PyYAML==6.0.2
albumentations==1.4.8

# --- Tipado / FastAPI deps ---
pydantic==2.8.2

